backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.1
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.5.0
PyMuPDF==1.26.3
python-docx==1.2.0
pytesseract==0.3.10
Pillow==11.3.0
openai==1.92.2
requests==2.32.4
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.1.0
python-jose[cryptography]==3.5.0